fixThreadsBelowFilter: fix domain used for threadIdxxScheduleDepth
threadIdxxScheduleDepthState is the mapping from active domain points to
the schedule depth of the band member mapped to thread x.  Since the
mapping happens below the given node in fixThreadsBelowFilter, we need
to use domain points active in a child of the given node rather than in
the given node itself (especially the given node being a filter).  Note
that the same domain points are active in all children of a node as only
ancestor filters are taken into account.  Take active domain points of
the first child of the given node instead of the node itself.

Without this change, threadIdxxScheduleDepthState could have multiple
entries for partially overlapping domains, potentially with different
values of depth.  However, memory promotion code assumes the depth must
be the same.  Furthermore, having different depths considered as mapped
to thread x for the same statement instance is incorrect with respect to
the parallel execution model.